requirements.txt
absl-py==1.4.0
altair==4.2.2
astunparse==1.6.3
attrs==22.2.0
blinker==1.6
blosc2==2.0.0
cachetools==5.3.0
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
decorator==5.1.1
dlclibrary==0.0.2
entrypoints==0.4
ffmpeg-python==0.2.0
filelock==3.11.0
filterpy==1.4.5
flatbuffers==23.3.3
fonttools==4.39.3
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
imgaug==0.4.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
llvmlite==0.39.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.0.4
msgpack==1.0.5
msgpack-numpy==0.4.8
networkx==3.1
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.5.0
plotly==5.14.1
protobuf==3.20.3
psutil==5.9.4
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
shapely==2.0.1
six==1.16.0
smmap==5.0.0
statsmodels==0.13.5
stqdm==0.0.5
streamlit==1.21.0
streamlit-option-menu==0.3.2
streamlit-authenticator==0.2.1
tables==3.8.0
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorpack==0.11
termcolor==2.2.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2023.3.21
toml==0.10.2
toolz==0.12.0
torch==1.12.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
